Refactoring db access, Rename Converter to CrudOperator and add operations, add FileService
