+++ src/AasxServerDB/Edit.cs
@@ -56,95 +56,5 @@ namespace AasxServerDB
             env.setWrite(false);
             Console.WriteLine("SAVE AASX TO DB: " + env.Filename);
         }
-
-        public static void DeleteAAS(string aasIdentifier)
-        {
-            using (AasContext db = new AasContext())
-            {
-                // Deletes automatically from DB
-                db.AASSets
-                    .Include(aas => aas.SMRefSets)
-                    .Where(aas => aas.Identifier == aasIdentifier).ExecuteDelete();
-
-                db.SaveChanges();
-            }
-        }
-
-        public static void DeleteSubmodel(string submodelIdentifier)
-        {
-            using (AasContext db = new AasContext())
-            {
-                using (var transaction = db.Database.BeginTransaction())
-                {
-                    try
-                    {
-                        var smDB = db.SMSets.Where(sm => sm.Identifier == submodelIdentifier);
-                        var smDBID = smDB.FirstOrDefault().Id;
-                        var smeDB = db.SMESets.Where(sme => sme.SMId == smDBID);
-                        var smeDBIDList = smeDB.Select(sme => sme.Id).ToList();
-
-                        db.SValueSets.Where(s => smeDBIDList.Contains(s.SMEId)).ExecuteDelete();
-                        db.IValueSets.Where(i => smeDBIDList.Contains(i.SMEId)).ExecuteDelete();
-                        db.DValueSets.Where(d => smeDBIDList.Contains(d.SMEId)).ExecuteDelete();
-                        db.OValueSets.Where(o => smeDBIDList.Contains(o.SMEId)).ExecuteDelete();
-                        smeDB.ExecuteDelete();
-                        smDB.ExecuteDelete();
-
-                        db.SaveChanges();
-                        transaction.Commit();
-                    }
-                    catch (Exception)
-                    {
-                        transaction.Rollback();
-                    }
-                }
-            }
-        }
-
-        public static void DeleteSubmodelElement(SMESet sME)
-        {
-            using (AasContext db = new AasContext())
-            {
-                using (var transaction = db.Database.BeginTransaction())
-                {
-                    try
-                    {
-                        db.SValueSets.Where(s => s.SMEId == sME.Id).ExecuteDelete();
-                        db.IValueSets.Where(i => i.SMEId == sME.Id).ExecuteDelete();
-                        db.DValueSets.Where(d => d.SMEId == sME.Id).ExecuteDelete();
-                        db.OValueSets.Where(o => o.SMEId == sME.Id).ExecuteDelete();
-                        db.SMESets.Where(sme => sme.Id == sME.Id).ExecuteDelete();
-
-                        db.SaveChanges();
-                        transaction.Commit();
-                    }
-                    catch (Exception)
-                    {
-                        transaction.Rollback();
-                    }
-                }
-            }
-        }
-
-        public static void DeleteConceptDescription(string conceptDescriptionId)
-        {
-            using (AasContext db = new AasContext())
-            {
-                using (var transaction = db.Database.BeginTransaction())
-                {
-                    try
-                    {
-                        db.CDSets.Where(cd => cd.Identifier == conceptDescriptionId).ExecuteDelete();
-
-                        db.SaveChanges();
-                        transaction.Commit();
-                    }
-                    catch (Exception)
-                    {
-                        transaction.Rollback();
-                    }
-                }
-            }
-        }
     }
 }
