+++ src/AasxServerStandardBib/SecurityClient.cs
@@ -3028,7 +3028,7 @@ namespace AasxServer
 
         public static void setTimeStampValue(string submodelId, string path, DateTime timeStamp, string value = null)
         {
-            Converter.setTimeStampValue(submodelId, path, timeStamp, value);
+            CrudOperator.setTimeStampValue(submodelId, path, timeStamp, value);
         }
         //public static AdminShellPackageEnv getEnv(string aasID)
         //{
